+++ packages/teamscale-coverage-collector/CHANGELOG.md
@@ -2,6 +2,10 @@ We use [Semantic Versioning](https://semver.org/).
 
 # New Release
 
+# 0.0.1-beta.30
+
+- [fix] The logger was logging asynchronously making it hard to debug the progress.
+
 # 0.0.1-beta.28
 
 - [fix] The logger could not resolve a dependency ('../diagnostics')