+++ packages/teamscale-javascript-instrumenter/CHANGELOG.md
@@ -2,6 +2,11 @@ We use [Semantic Versioning](https://semver.org/).
 
 # New Release
 
+# 0.0.1-beta.30
+
+- [fix] The logger was logging asynchronously making it hard to debug the progress.
+- [fix] Too many files were instrumented in parallel causing memory and performance issues.
+
 # 0.0.1-beta.29
 
 - [fix] Branch counters were present even though they are currently not supported.